pyproject.toml
[tool.poetry]
name = "hw7"
version = "0.1.0"
description = ""
authors = ["Anton Holovin <54354764+Unfeir@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.4"
psycopg2-binary = "^2.9.5"
alembic = "^1.9.4"
faker = "^17.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
